WW-4034 Refactors and extracts interface from JSONWriter
