+++ plugins/json/src/test/java/org/apache/struts2/json/JSONEnumTest.java
@@ -46,7 +46,7 @@ public class JSONEnumTest extends TestCase {
         bean1.setEnumField(AnEnum.ValueA);
         bean1.setEnumBean(AnEnumBean.Two);
 
-        JSONWriter jsonWriter = new JSONWriter();
+        JSONWriter jsonWriter = new DefaultJSONWriter();
         jsonWriter.setEnumAsBean(false);
         String json = jsonWriter.write(bean1);
 
@@ -88,7 +88,7 @@ public class JSONEnumTest extends TestCase {
         bean1.setEnumField(AnEnum.ValueA);
         bean1.setEnumBean(AnEnumBean.Two);
 
-        JSONWriter jsonWriter = new JSONWriter();
+        JSONWriter jsonWriter = new DefaultJSONWriter();
         jsonWriter.setEnumAsBean(true);
         String json = jsonWriter.write(bean1);
